requirements.txt
python-telegram-bot==20.8
instagrapi==2.1.2
apscheduler==3.10.4
Flask==3.0.0
